// frontend/apps/web/src/domains/api/generated/model/jobStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Applark
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const JobStatus = {
  pending: 'pending',
  scraping: 'scraping',
  extracting: 'extracting',
  ready: 'ready',
  failed: 'failed',
} as const;